backend/requirements.txt

Django>=5.2.2
djangorestframework>=3.16.0
django-filter>=25.2
django-cors-headers>=4.9.0
pymysql